chore: remove deprecated collab integration and fix IDOR in generateIssueContent
- Delete collab-task-action.ts and create-collab-task-dialog.tsx
- Move generateIssueContent to shared issue-content-action.ts with org ownership check
- Update GitHub/Linear dialog imports to use new shared file
- Remove collabConnected prop and collab queries from issues page and dashboard
- Clean up collab references from issues-content and recent-issues components

Closes #216
